[loading screen] moved observable logic code to util file
